Add supporting files to run components from distributed binaries (#552)
* Add supporting files

* Add EF tools

+++ builds/make-release.ps1
@@ -1,6 +1,7 @@
 param (
     [string]$version = "",
-    [string]$runtime = "win-x64"
+    [string]$runtime = "win-x64",
+    [switch]$noCompress = $false
 )
 
 if ($version -eq "") {
@@ -8,74 +9,55 @@ if ($version -eq "") {
 }
 
 $rootPath = Split-Path $PSScriptRoot
-$publishOuterPath = Join-Path $rootPath "/release"
+$publishOuterPath = Join-Path $rootPath "/release/opencatapult-v$version-$runtime"
 
 function Publish-Component {
     Param(
         [parameter(Position=1)][string]$csprojPath,
         [parameter(Position=2)][string]$publishPath,
-        [parameter(Position=3)][switch]$isTaskProvider = $false
+        [parameter(Position=3)][switch]$isTaskProvider = $false,
+        [parameter(Position=4)][switch]$copyCert = $false
     )
 
     Write-Host "Publishing component..."
 
     dotnet publish $csprojPath -c Release -o $publishPath -r $runtime --self-contained false /p:Version=$version
-    
-    Write-Host "Compressing the package..."
 
-    $compressSrcPath = "$publishPath/*"
-    $compressDestPath = "$publishPath-v$version-$runtime.zip"
+    Remove-Item "$publishPath/*" -Include appsettings.*.json
 
-    if ($isTaskProvider) {
-        $compressSrcPath = Join-Path $publishOuterPath "/plugins/*"
-        $taskProviderName = Split-Path $publishPath -Leaf
-        $compressDestPath = Join-Path $publishOuterPath "$taskProviderName-v$version-$runtime.zip"
+    if ($copyCert) {
+        $certPath = Join-Path $rootPath "/tools/certs/opencatapultlocal.pfx"
+
+        Copy-Item -Path $certPath -Destination $publishPath
     }
-
-    Compress-Archive -Path $compressSrcPath -DestinationPath $compressDestPath -Force
     
-    Write-Host "Deleting temporary files.."
+    if (!$noCompress) {
+        Write-Host "Compressing the package..."
 
-    $removeItemPath = $publishPath
-
-    if ($isTaskProvider) {
-        $removeItemPath = Join-Path $publishOuterPath "/plugins"
-    }
-
-    Remove-Item -Path $removeItemPath -Recurse -Force
-
-    Write-Host "Done. The package has been published to $compressDestPath." -ForegroundColor Green
-}
-
-function Publish-Web {
-    Param(
-        [parameter(Position=1)][string]$projectPath,
-        [parameter(Position=2)][string]$publishPath
-    )
-
-    Write-Host "Publishing web component..."
-
-    $distPath = Join-Path $projectPath "/dist"
-
-    Write-Host "npm install $projectPath"
-    npm install $projectPath
-    
-    Write-Host "npm run build-prod -- --outputPath=$distPath"
-    Set-Location $projectPath
-    npm run build-prod -- --outputPath=$distPath
-    Set-Location $rootPath
-
-    if (Test-Path $distPath) {
-        $compressSrcPath = "$distPath/*"
+        $compressSrcPath = "$publishPath/*"
         $compressDestPath = "$publishPath-v$version-$runtime.zip"
 
-        Compress-Archive -Path $compressSrcPath -DestinationPath $compressDestPath -Force
+        if ($isTaskProvider) {
+            $compressSrcPath = Join-Path $publishOuterPath "/plugins/*"
+            $taskProviderName = Split-Path $publishPath -Leaf
+            $compressDestPath = Join-Path $publishOuterPath "$taskProviderName-v$version-$runtime.zip"
+        }
 
+        Compress-Archive -Path $compressSrcPath -DestinationPath $compressDestPath -Force
+        
         Write-Host "Deleting temporary files.."
 
-        Remove-Item -Path $distPath -Recurse -Force
+        $removeItemPath = $publishPath
 
-        Write-Host "Done. The web package has been published to $compressDestPath." -ForegroundColor Green
+        if ($isTaskProvider) {
+            $removeItemPath = Join-Path $publishOuterPath "/plugins"
+        }
+
+        Remove-Item -Path $removeItemPath -Recurse -Force
+
+        Write-Host "Done. The package has been published to $compressDestPath." -ForegroundColor Green
+    } else {
+        Write-Host "Done. The component has been published to $publishPath." -ForegroundColor Green
     }
 }
 
@@ -85,7 +67,7 @@ Write-Host
 Write-Host "1. Processing API component" -ForegroundColor Yellow
 $apiCsprojPath = Join-Path $rootPath "/src/API/Polyrific.Catapult.Api/Polyrific.Catapult.Api.csproj"
 $apiPublishPath = Join-Path $publishOuterPath "/api"
-Publish-Component $apiCsprojPath $apiPublishPath
+Publish-Component $apiCsprojPath $apiPublishPath $false $true
 Write-Host ""
 
 Write-Host "2. Processing CLI component" -ForegroundColor Yellow
@@ -101,9 +83,9 @@ Publish-Component $engineCsprojPath $enginePublishPath
 Write-Host ""
 
 Write-Host "4. Processing Web component" -ForegroundColor Yellow
-$webProjectPath = Join-Path $rootPath "/src/Web/opencatapultweb"
+$webProjectPath = Join-Path $rootPath "/src/Web/Polyrific.Catapult.Web/Polyrific.Catapult.Web.csproj"
 $webPublishPath = Join-Path $publishOuterPath "/web"
-Publish-Web $webProjectPath $webPublishPath
+Publish-Component $webProjectPath $webPublishPath $false $true
 Write-Host ""
 
 Write-Host "5. Processing Task Provider components" -ForegroundColor Yellow
@@ -133,5 +115,30 @@ $plugins = [System.Tuple]::Create($aspNetCoreMvcCsprojPath, $aspNetCoreMvcPublis
 
 foreach ($p in $plugins) {
     Publish-Component $p.Item1 $p.Item2 $true
+    Write-Host ""
+}
+
+if ($noCompress) {
+
+    $pluginsFolder = Join-Path $publishOuterPath "/plugins"
+    Move-Item "$pluginsFolder" -Destination "$enginePublishPath"
+
+    Write-Host "6. Copying additional files" -ForegroundColor Yellow
+
+    $certs = Join-Path $rootPath "/tools/certs"
+    $dest = Join-Path $publishOuterPath "/certs"
+    Copy-Item "$certs" -Destination "$dest" -Recurse
+
+    $runScripts = Join-Path $rootPath "/tools/scripts/run*"
+    Copy-Item "$runScripts" -Destination "$publishOuterPath"
+
+    $efDll = Join-Path $rootPath "/tools/ef.dll"
+    $dest = Join-Path $publishOuterPath "/tools"
+    if (!(Test-Path $dest)) {
+        New-Item $dest -ItemType Directory | Out-Null
+    }
+    Copy-Item "$efDll" -Destination "$dest"
+
+    Write-Host "Done." -ForegroundColor Green
     Write-Host ""
 }
